+++ oidc-stub/src/utils/EnvironmentVarUtils.ts
@@ -1,0 +1,8 @@
+export function checkRequiredEnvVars(requiredVars: string[]): boolean {
+	const missingVars = requiredVars.filter(varName => !process.env[varName]);
+	if (missingVars.length > 0) {
+		return false;
+	}
+	return true;
+}
+
